+++ awesomeNations/main.py
@@ -1,4 +1,6 @@
-from awesomeNations.objects import NationObject, RegionObject
+from awesomeNations.nationMethods import N
+from awesomeNations.regionMethods import R
+from typing import Iterator
 from datetime import datetime
 
 class AwesomeNations:
@@ -18,6 +20,9 @@ class AwesomeNations:
         age = today - created
         result = f'Around {age-1}-{age} years old.'
         return result
+    
+    def configure(export = None):
+        pass
 
     class Nation:
         """
@@ -30,24 +35,24 @@ class AwesomeNations:
             """
             Checks if nation exists.
             """
-            exist = NationObject.exists(self)
+            exist = N(self.nation_name).exists()
             return exist
         
         def get_overview(self) -> dict:
             """
             Get an overview of the requested nation.
             """
-            overview = NationObject.overview(self)
+            overview = N(self.nation_name).overview()
             return overview
 
-        def get_census(self, censusid: list = [0]) -> dict:
+        def get_census(self, censusid: tuple = [0]) -> Iterator:
             """
             Gets one or more censuses [0-88] from the requested nation, examples:
             - [0]: Civil rights
             - [46]: Defense Forces
             - [0, 1, 2, 3] ... [88]
             """
-            census = NationObject.census(self, censusid)
+            census = N(self.nation_name).census_generator(censusid)
             return census
 
     class Region:
@@ -61,28 +66,28 @@ class AwesomeNations:
             """
             Checks if the region exists.
             """
-            exist = RegionObject.exists(self)
+            exist = R.exists(self)
             return exist
         
         def get_overview(self) -> dict:
             """
             Provides an overview of the requested region.
             """
-            overview = RegionObject.overview(self)
+            overview = R.overview(self)
             return overview
 
         def get_world_census(self, censusid: int) -> dict:
             """
             Retrieves the world census rankings for the requested region.
             """
-            ranks = RegionObject.world_census(self, censusid)
+            ranks = R.world_census(self, censusid)
             return ranks
 
         def get_embassies(self) -> dict:
             """
             Retrieves the embassies of the requested region.
             """
-            embassies = RegionObject.embassies(self)
+            embassies = R.embassies(self)
             return embassies
 
 
@@ -94,9 +99,16 @@ if __name__ == '__main__':
     print('NATION')
     print(f'{AwesomeNations.Nation().exists()=}')
     print(f'{AwesomeNations.Nation().get_overview()=}')
-    print(f'{AwesomeNations.Nation().get_census()=}')
+    # print(f'{AwesomeNations.Nation().get_census()=}')
+
+    for stuff in AwesomeNations.Nation().get_census((i for i in range(89))):
+        print(stuff)
 
     print('REGION')
     print(f'{AwesomeNations.Region().exists()=}')
     print(f'{AwesomeNations.Region().get_overview()=}')
     print(f'{AwesomeNations.Region().get_world_census(46)=}')
+    # print(f'{AwesomeNations.Region().get_embassies()=}')
+
+    for stuff in AwesomeNations.Region().get_embassies():
+        print(stuff)
